+++ src/software/sensor_fusion/filter/ball_filter_test.cpp
@@ -114,8 +114,8 @@ class BallFilterTest : public ::testing::Test
         // time step in order for the ball to reach the end of the given segment.
         Duration max_ball_travel_duration =
             Duration::fromSeconds(ball_path.length() / ball_velocity_magnitude);
-        int num_iterations =
-            std::round(max_ball_travel_duration.getSeconds() / time_step.getSeconds());
+        int num_iterations = static_cast<int>(
+            std::round(max_ball_travel_duration.getSeconds() / time_step.getSeconds()));
 
         testFilterHelper(start_time, ball_starting_position, ball_velocity,
                          ball_position_variance, time_step_variance,
